requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.6.0
async-generator==1.10
async-timeout==4.0.2
attrs==22.1.0
beautifulsoup4==4.11.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==39.0.1
distlib==0.3.6
Django==4.1.5
django-cors-headers==3.13.0
django_debug_toolbar==3.8.1
djangorestframework==3.14.0
dnspython==2.2.1
filelock==3.8.0
funcy==1.18
gunicorn==20.1.0
h11==0.14.0
idna==3.4
mysql-connector-python==8.0.32
numpy==1.23.5
outcome==1.2.0
pandas==1.5.2
Pillow==9.4.0
pipenv==2022.11.11
platformdirs==2.5.4
protobuf==3.20.3
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
pymemcache==4.0.0
pymongo==4.3.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-magic-bin==0.4.14
pytz==2022.6
redis==4.4.2
requests==2.28.1
scipy==1.10.1
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
stripe==5.2.0
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
urllib3==1.26.13
virtualenv==20.16.7
virtualenv-clone==0.5.7
wsproto==1.2.0
